:hocho: Refactor logger utils away from generic utils
